adding a new test for kdtrees

+++ src/vectors/array_backed.hh
@@ -116,7 +116,7 @@ namespace vectors {
   {
     os << "{ ";
     for (size_t i = 0; i < v.size (); ++i)
-      os << v[i] << " ";
+      os << (int) v[i] << " ";
     os << "}";
     return os;
   }
